# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AddAttendeeRequest:
    properties:
      userId:
        maxLength: 100
        type: string
    required:
    - userId
    type: object
...
  domain.CompleteActivityRequest:
    properties:
      outcome:
        maxLength: 500
        type: string
    type: object
...
  domain.CreateFollowUpRequest:
    properties:
      assignedToId:
        maxLength: 100
        type: string
      description:
        maxLength: 2000
        type: string
      dueDate:
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - title
    type: object
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  domain.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@straye.io
    name: API Support
  description: CRM Relation API for customer, project, and sales pipeline management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Straye Relation API
  version: "1.0"
paths:
...
  /activities:
    get:
      consumes:
      - application/json
      description: Get paginated list of activities with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 200)
        in: query
        name: pageSize
        type: integer
      - description: Filter by activity type (meeting, task, call, email, note)
        in: query
        name: type
        type: string
      - description: Filter by status (planned, in_progress, completed, cancelled)
        in: query
        name: status
        type: string
      - description: Filter by target type (customer, project, offer, deal)
        in: query
        name: targetType
        type: string
      - description: Filter by target entity ID
        format: uuid
        in: query
        name: targetId
        type: string
      - description: Filter by assigned user ID
        in: query
        name: assignedTo
        type: string
      - description: Filter activities from this date (YYYY-MM-DD), inclusive from
          00:00:00
        in: query
        name: from
        type: string
      - description: Filter activities to this date (YYYY-MM-DD), inclusive until
          23:59:59
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ActivityDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List activities
      tags:
      - Activities
    post:
      consumes:
      - application/json
      description: Create a new activity (meeting, task, call, email, or note)
      parameters:
      - description: Activity data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create activity
      tags:
      - Activities
  /activities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an activity
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete activity
      tags:
      - Activities
    get:
      consumes:
      - application/json
      description: Get a single activity by its ID
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get activity by ID
      tags:
      - Activities
    put:
      consumes:
      - application/json
      description: Update an existing activity
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Activity data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update activity
      tags:
      - Activities
  /activities/{id}/attendees:
    post:
      consumes:
      - application/json
      description: Add a user as an attendee to a meeting activity
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Attendee user ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.AddAttendeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Add attendee to meeting
      tags:
      - Activities
  /activities/{id}/attendees/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a user from the attendees list of a meeting activity
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Remove attendee from meeting
      tags:
      - Activities
  /activities/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark an activity as completed with an optional outcome
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Optional outcome
        in: body
        name: body
        schema:
          $ref: '#/definitions/domain.CompleteActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Complete activity
      tags:
      - Activities
  /activities/{id}/follow-up:
    post:
      consumes:
      - application/json
      description: Create a follow-up task from a completed activity
      parameters:
      - description: Activity ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Follow-up task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateFollowUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ActivityDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create follow-up task
      tags:
      - Activities
  /activities/my-tasks:
    get:
      consumes:
      - application/json
      description: Get current user's incomplete tasks (assigned tasks that are not
        completed or cancelled)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 200)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ActivityDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get my tasks
      tags:
      - Activities
  /activities/stats:
    get:
      consumes:
      - application/json
      description: Get activity status counts for the current user's dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ActivityStatusCounts'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get activity statistics
      tags:
      - Activities
  /activities/upcoming:
    get:
      consumes:
      - application/json
      description: Get upcoming scheduled activities for the current user within a
        specified number of days
      parameters:
      - default: 7
        description: Number of days to look ahead (1-90)
        in: query
        name: daysAhead
        type: integer
      - default: 20
        description: Maximum number of activities to return (1-100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ActivityDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get upcoming activities
      tags:
      - Activities
  /audit:
    get:
      description: Returns a paginated list of audit log entries with optional filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by action type
        in: query
        name: action
        type: string
      - description: Filter by entity type
        in: query
        name: entityType
        type: string
      - description: Filter by entity ID
        in: query
        name: entityId
        type: string
      - description: Filter by start time (RFC3339)
        in: query
        name: startTime
        type: string
      - description: Filter by end time (RFC3339)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuditLogListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List audit logs
      tags:
      - Audit
  /audit/{id}:
    get:
      description: Returns a specific audit log entry
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuditLogDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get audit log by ID
      tags:
      - Audit
...
  /companies:
    get:
      description: Returns a list of all active Straye group companies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Company'
            type: array
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get all companies
      tags:
      - Companies
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for system operations
    in: header
    name: x-api-key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"